fix: floating point division

+++ VectorManipulation.cpp
@@ -9,7 +9,7 @@ long long getSum(const std::vector<long long>& numbers) {
     return sum;
 }
 
-long long getAvg(const std::vector<long long>& numbers) {
+double getAvg(const std::vector<long long>& numbers) {
     long long sum = getSum(numbers);
     return 1.0 * sum / numbers.size();
 }
